show data in recycler

+++ app/src/main/java/com/cvte/taobaounion/model/domain/HomePagerContent.java
@@ -316,7 +316,7 @@ public class HomePagerContent {
             return volume;
         }
 
-        public void setVolume(long volume) {
+        public void setVolume(int volume) {
             this.volume = volume;
         }
